third_party/WebKit/LayoutTests/fast/events/platform-wheelevent-paging-xy-in-scrolling-page-expected.txt

PASS event.wheelDeltaY is window.givenScrollTop * -3
PASS event.wheelDeltaX is window.givenScrollLeft * -3
PASS event.wheelDelta is window.givenScrollTop * -3
PASS document.scrollingElement.scrollTop == window.expectedScrollTop && document.scrollingElement.scrollLeft == window.expectedScrollLeft became true
PASS successfullyParsed is true

TEST COMPLETE
